+++ src/components/admin/ProductVariantMapper.php
@@ -31,16 +31,18 @@ readonly class ProductVariantMapper
         $imageId = $this->data['media']['nodes'][0]['id'] ?? null;
         $this->variant->image_id = $imageId ? (new ShopifyId($imageId))->toInt() : null;
 
-        $this->variant->barcode = $this->data['barcode'] ?: null;
-        $this->variant->compare_at_price = $this->data['compareAtPrice'];
+        $this->variant->price = (new ShopifyPrice($this->data['price']))->toInt();
+        $this->variant->compare_at_price = $this->data['compareAtPrice']
+            ? (new ShopifyPrice($this->data['compareAtPrice']))->toInt()
+            : null;
 
+        $this->variant->name = $this->data['title'];
+        $this->variant->sku = $this->data['sku'] ?: null;
+        $this->variant->barcode = $this->data['barcode'] ?: null;
         $this->variant->inventory_policy = $this->data['inventoryPolicy'] ?? null;
         $this->variant->inventory_quantity = $this->data['inventoryQuantity'];
         $this->variant->inventory_tracked = $this->data['inventoryItem']['tracked'];
         $this->variant->is_taxable = $this->data['taxable'] ?? false;
-        $this->variant->name = $this->data['title'];
-        $this->variant->price = $this->data['price'];
-        $this->variant->sku = $this->data['sku'] ?: null;
         $this->variant->weight = $this->data['inventoryItem']['measurement']['weight']['value'] ?? null;
 
         $this->variant->weight_unit = match ($this->data['inventoryItem']['measurement']['weight']['unit'] ?? null) {